// src/client/types.gen.ts
export type ToDoListPublic = {
    /**
     * Title
     */
    title: string;
    /**
     * Description
     */
    description?: string | null;
    status?: ToDoStatus;
    /**
     * Due Date Time
     */
    due_date_time?: string | null;
    priority?: ToDoPriority;
    /**
     * Id
     */
    id: string;
    /**
     * User Id
     */
    user_id: string;
    /**
     * Created At
     */
    created_at: string;
};
// ...
export type ToDoPriority = 'high' | 'medium' | 'low';

/**
 * ToDoStatus
 */
export type ToDoStatus = 'pending' | 'in progress' | 'completed';